==> scripts/0_inputs.R <==
## any inputs that need to be changed from year to year

post_16_school_census_filename <- "1516_2223_ncy.csv"

final_school_period <- "2223"

max_year <- 2023 # the later year in the final school period

max_output_year <- max_year + 9
